% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-documentation.R
\name{my_func}
\alias{my_func}
\title{Sample R function}
\usage{
my_func(x, y)
}
\arguments{
\item{x, y}{\code{numeric} Data vectors}
}
\value{
A single \code{numeric} value
}
\description{
For this function, we...
}
\examples{

my_func(1:2, 3:4)

}
\author{
Blame this person
}
