% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmsapkg-package.R
\docType{package}
\name{tmsapkg-package}
\alias{tmsapkg}
\alias{tmsapkg-package}
\title{tmsapkg: Your Title Here}
\description{
Brief description...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmsalab/$PACKAGE$}
  \item Report bugs at \url{https://github.com/tmsalab/$PACKAGE$/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{NetID@illinois.edu} (10110101) [copyright holder]

Authors:
\itemize{
  \item First2 Last3 \email{NetID@illinois.edu} (0189) [copyright holder]
}

}
